Crate gazebo

A collection of well-tested primitives that have been useful. Most modules stand alone.

Modules

cast

Cast between types with no conversion.

cell

Additions to the Ref mechanism.

cmp

Traits to help implementing dynamic comparisons.

file

File/IO operations.

hash

Utilities for working with hashes.

phantom

Additional PhantomData related types.

prelude

Standard functions. Usually imported with use gazebo::prelude::*.

types

Operations working on Rust types.

variants

Working with the variants of an enum.

Macros

cmp_chain

Performs a chain of comparison operation expressions yielding std::cmp::Ordering, supporting early exit upon hitting the first expressions that doesn’t yield std::cmp::Ordering::Equal and returning the result of that. This is useful for easily writing a sequence of expressions necessary to yield a comparison result. The macro is expanded inplace, so any expressions dealing with Result types are allowed provided that the larger scope allows returning result.

eq_chain

Performs a chain of equals operation expressions yielding bool, supporting early exit upon hitting the first expressions that returns false and returning false. This is useful for easily writing a sequence of equals expressions necessary to yield a bool The macro is expanded inplace, so any expressions dealing with Result types are allowed provided that the larger scope allows returning result.

transmute

transmute!(from-type, to-type, value) will do a transmute, but the original and result types must be specified.