gazebo::cast

Function transmute_unchecked

pub unsafe fn transmute_unchecked<A, B>(x: A) -> B

Like normal transmute, but without the compile-time check that the sizes of the input and output are the same. Despite the removal of the compile-time check, this property must still hold, and there is a debug_assert_eq which checks that. All the rules and warnings from transmute still apply.

This function is mostly useful when you have types that are polymorphic, e.g. Vec<T>, that transmute cannot be applied to.