gazebo::cell

Trait AsARef

pub trait AsARef<T: ?Sized> {
    // Required methods
    fn as_aref(this: &Self) -> ARef<'_, T>;
    fn try_as_aref(this: &Self) -> Result<ARef<'_, T>, BorrowError>;
    fn as_ref_cell(this: &Self) -> Option<&RefCell<T>>;
}

Obtain an ARef from either a normal pointer or a RefCell.

Required Methods

fn as_aref(this: &Self) -> ARef<'_, T>

Get an ARef pointing at this type.

fn try_as_aref(this: &Self) -> Result<ARef<'_, T>, BorrowError>

Try and get an ARef pointing at this type. Returns an Err if the type Self is a RefCell which is already mutably borrowed.

fn as_ref_cell(this: &Self) -> Option<&RefCell<T>>

Return the underlying RefCell if Self is one, otherwise None.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.