gazebo

Module prelude

Standard functions. Usually imported with use gazebo::prelude::*.

Contains:

Extension methods for str and slice/Vec.

Defines Default_ macro.

The derivation macros appended with underscore are like the normal derivations, but don’t require the trait on any argument types. For example, given the type:

#[derive(Default_)]
struct Foo<T>(std::marker::PhantomData<T>);

Traits

IterExt

Extension traits on Iterator.

IterOwned

OptionExt

Extension traits on Option where it holds any value or ref.

SliceClonedExt

Short hand analogous to Iter::cloned, where items of &T are converted to T via clone.

SliceCopiedExt

Short hand analogous to Iter::copied, where items of &T are converted to T via copy.

SliceDupedExt

Short hand analogous to Iter::duped, where items of &T are converted to T via dupe.

SliceExt

Extension traits on slices/Vec.

VecExt

Extension traits on Vec.

Derive Macros

Default_

Derive the Default trait, but without requiring all type arguments to implement [Default].