gazebo::types

Trait TEq

pub trait TEq<T> {
    // Required methods
    fn teq(self) -> T;
    fn teq_ref(&self) -> &T;
    fn teq_mut(&mut self) -> &mut T;
}

A trait witnessing that two types are equal.

For example:

use gazebo::types::TEq;
fn foo<A: TEq<String>>(x: A) -> String {
    x.teq()
}

You should not make any further implementations of this trait.

Originally taken from a request to have Rust allow equality constraints in traits, from Issue 20041.

Required Methods

fn teq(self) -> T

Convert between two equal types.

fn teq_ref(&self) -> &T

Convert between references to two equal types.

fn teq_mut(&mut self) -> &mut T

Convert between mutable references to two equal types.