gazebo

Module variants

Working with the variants of an enum.

Traits

VariantName

Derive Macros

UnpackVariants

Generates implementation to unpack the inner data of enum variants. The unpack_XXX methods return the unpacked data as a tuple of references of the inner data. The into_XXX variants return similarly-structured tuples.

VariantName

Trait for enums to return the name of the current variant as a str. Useful for debugging messages.