gazebo::cell

Struct ARef

pub struct ARef<'a, T: ?Sized + 'a>(/* private fields */);

A Ref that might not actually be borrowed. Either a Ptr (a normal & style reference), or a Ref (like from RefCell), but exposes all the methods available on Ref.

Implementations

impl<'a, T: ?Sized + 'a> ARef<'a, T>

pub fn new_ptr(x: &'a T) -> Self

Create a new ARef from a pointer.

pub fn new_ref(x: Ref<'a, T>) -> Self

Create a new ARef from a reference.

pub fn clone(orig: &Self) -> Self

See Ref.clone. Not a self method since that interferes with the Deref.

pub fn map<U: ?Sized, F>(orig: ARef<'a, T>, f: F) -> ARef<'a, U>
where F: FnOnce(&T) -> &U,

See Ref.map. Not a self method since that interferes with the Deref.

pub fn map_split<U: ?Sized, V: ?Sized, F>( orig: ARef<'a, T>, f: F, ) -> (ARef<'a, U>, ARef<'a, V>)
where F: FnOnce(&T) -> (&U, &V),

See Ref.map_split. Not a self method since that interferes with the Deref.

pub fn filter_map<U: ?Sized, F>( orig: ARef<'a, T>, f: F, ) -> Result<ARef<'a, U>, Self>
where F: FnOnce(&T) -> Option<&U>,

See Ref.filter_map. Not a self method since that interferes with the Deref.